first run at apollofindrx tests
